refactor: 이름 조건 검증 기능 위치 변경
